+++ drift_flutter/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.4-dev
 
 - Allow providing a custom temporary directory.
+- Allow providing a custom database directory, making it easier to swap out the
+  default `getApplicationDocumentsDirectory()`.
 
 ## 0.2.3